device-orientation-look: Move to TypeScript
